Fix return type for programInput and previewInput

+++ qatemconnection.h
@@ -119,9 +119,9 @@ public:
     bool debugEnabled() const { return m_debugEnabled; }
 
     /// @returns the index of the input that is on program
-    quint8 programInput() const { return m_programInput; }
+    quint16 programInput() const { return m_programInput; }
     /// @returns the index of the input that is on preview
-    quint8 previewInput() const { return m_previewInput; }
+    quint16 previewInput() const { return m_previewInput; }
 
     /// @returns the tally state of the input @p id. 1 = program, 2 = preview and 3 = both
     quint8 tallyState(quint8 id) const;
